Converted "love" to "defined" variable
