+++ src/game/load.lua
@@ -1,0 +1,9 @@
+local function init(love)
+    local function load()
+
+    end
+
+    return load
+end
+
+return init
