+++ src/game/update.lua
@@ -1,0 +1,9 @@
+local function init(love)
+    local function update(dt)
+
+    end
+
+    return update
+end
+
+return init
