+++ src/input/keyboard.lua
@@ -1,14 +1,18 @@
-local function keypressed(key)
+local function init(love)
+    local function keypressed(key)
 
+    end
+
+
+    local function keyreleased(key)
+
+    end
+
+
+    return {
+        keypressed = keypressed,
+        keyreleased = keyreleased,
+    }
 end
 
-
-local function keyreleased(key)
-
-end
-
-
-return {
-    keypressed = keypressed,
-    keyreleased = keyreleased,
-}
+return init
