+++ src/input/mouse.lua
@@ -1,14 +1,18 @@
-local function mousepressed(x, y, button, istouch)
+local function init(love)
+    local function mousepressed(x, y, button, istouch)
 
+    end
+
+
+    local function mousereleased(x, y, button, istouch)
+
+    end
+
+
+    return {
+        mousepressed = mousepressed,
+        mousereleased = mousereleased,
+    }
 end
 
-
-local function mousereleased(x, y, button, istouch)
-
-end
-
-
-return {
-    mousepressed = mousepressed,
-    mousereleased = mousereleased,
-}
+return init
