+++ src/input/window.lua
@@ -1,14 +1,18 @@
-local function focus(f)
+local function init(love)
+    local function focus(f)
 
+    end
+
+
+    local function quit()
+
+    end
+
+
+    return {
+        focus = focus,
+        quit = quit,
+    }
 end
 
-
-local function quit()
-
-end
-
-
-return {
-    focus = focus,
-    quit = quit,
-}
+return init
